pyproject.toml
[tool.pytest.ini_options]
# относительно root(папки с проектом)
# добавляются директории
# . - текущая директория root (отсутствует при pytest)
# src - директория с исходниками (тоже отсутствует при pytest)
# присутствует только директория tests
pythonpath = [".", "src"]
asyncio_mode="auto" # auto, strict

[tool.poetry]
name = "shift-fastapi-project"
version = "0.1.0"
description = "My fast-api application"
authors = ["Daniel <d.akipskiy@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "shift_fastapi_project"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
fastapi = "^0.97.0"
fastapi-users-db-sqlalchemy = "^5.0.0"
sqlalchemy = "^2.0.16"
alembic = {extras = ["sqlalchemy"], version = "^1.11.1"}
fastapi-users = {extras = ["redis"], version = "^11.0.0"}
httpx = "^0.24.1"
asyncpg = "^0.27.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
gunicorn = "^19"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
aioredis = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-asyncio = "^0.21.0"
uvicorn = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
